[SUPPORT] Simplify BottomSheet implementation
