+++ design/preview/src/main/java/com/mikhaellopez/designsystem/DesignSystemActivity.kt
@@ -3,7 +3,6 @@ package com.mikhaellopez.designsystem
 import android.os.Bundle
 import androidx.activity.compose.setContent
 import androidx.appcompat.app.AppCompatActivity
-import com.mikhaellopez.ui.base.getViewGroupParent
 import com.mikhaellopez.ui.theme.BaseTheme
 
 class DesignSystemActivity : AppCompatActivity() {
@@ -12,7 +11,7 @@ class DesignSystemActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContent {
             BaseTheme {
-                DesignSystem(viewGroupParent = getViewGroupParent())
+                DesignSystem()
             }
         }
     }
